requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozendict==2.4.6
html5lib==1.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
multitasking==0.0.11
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
soupsieve==2.6
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.67.1
tweepy==4.14.0
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.51
feedparser==6.0.11
fastapi==0.115.8
uvicorn==0.30.6
vaderSentiment==3.3.2
